Refactor: move out integration pipeline code and rename helpers
Slightly reorder of the actions in main.go, but should not have an
impact on actual behaviour.
