+++ menderqa_pipeline_helpers.go
